feat(table-plugin): new lightweight API for react-native-render-html v5.0
This release supports the new release of react-native-render-html!

BREAKING CHANGE: `makeTableRenderer` and `makeCustomTableRenderer` have
been dropped in favor of the default `table` renderer export and
`extractHtmlTableProp` function for custom renderers. This release takes
advantage of the availability of `domNode` in custom renderers and the
new `domNodeToHTMLString` utility available in react-native-render-html.
Configuration for the table renderer is now read from
`renderersProps.table` prop of the `HTML` component.